Add init script to run commands on reboot
After a dist-upgrade, there are scenarios where we
still need to perform some operations to leave
uaclient on a desired state. For example, from trusty
to xenial, we need to try to enable livepatch. However,
we can only do that after the user has upgraded the machine
and rebooted it. To handle those situations, we are now
creating a reboot init script that should run after the user
reboot the machine for the first time and finish the remaining
uaclient operations that could not be performed on upgrade.